nat: extract counter-store/clear tests into tests_counter.rs (#4409)
Move the 5 NAT hit-counter tests -- nat_counter_store_* (id-zero skip,
stable ids across reorder/removal), parsed_nat_rules_share_store_counters
and the two #3830 nat_counter_clear_* concurrency tests -- verbatim into
tests_counter.rs. These carry no test-local helpers.

Pure code motion: no test body edited, nothing renamed. nat/tests.rs
now holds 41 tests.
